修复时间戳计算中的数据类型溢出问题
- 将calculateTimestamp返回类型从uint32_t改为uint64_t
- 修复NTP偏移量(1.7万亿ms)超出uint32_t范围导致的溢出问题
- 更新formatTimestamp方法参数类型为uint64_t
- 更新UartReceiver中的调用以使用新的返回类型
- 解决时间戳计算错误的问题，现在应该显示正确的北京时间

+++ include/TimeSync.h
@@ -37,13 +37,13 @@ public:
     void addTimePair(uint8_t sensorId, uint32_t sensorTimeMs, int64_t espTimeUs);
     
     // 计算时间戳：T = a*S + b + N（快速操作）
-    uint32_t calculateTimestamp(uint8_t sensorId, uint32_t sensorTimeMs);
+    uint64_t calculateTimestamp(uint8_t sensorId, uint32_t sensorTimeMs);
     
     // 后台拟合计算（在后台任务中调用）
     void performBackgroundFitting();
     
     // 格式化时间戳为时/分/秒/毫秒格式
-    uint32_t formatTimestamp(uint32_t timestampMs);
+    uint32_t formatTimestamp(uint64_t timestampMs);
     
     // 获取NTP时间差N（毫秒）
     int64_t getNtpOffset() const;
